+++ src/fstcore/byte/byte_v12.h
@@ -39,8 +39,9 @@
 
 #include <fstream>
 
-void fdsWriteByteVec_v12(std::ofstream &myfile, char* byteVector, unsigned int nrOfRows, unsigned int compression, std::string annotation);
+void fdsWriteByteVec_v12(std::ofstream &myfile, char* byteVector, unsigned long long nrOfRows, unsigned int compression, std::string annotation);
 
-void fdsReadByteVec_v12(std::istream &myfile, char* byteVector, unsigned long long blockPos, unsigned startRow, unsigned length, unsigned size);
+void fdsReadByteVec_v12(std::istream &myfile, char* byteVector, unsigned long long blockPos, unsigned long long startRow,
+  unsigned long long length, unsigned long long size);
 
 #endif // BYTE_V12_H
