+++ src/fstcore/character/character_v6.h
@@ -47,7 +47,8 @@
 void fdsWriteCharVec_v6(std::ofstream &myfile, IStringWriter* blockRunner, int compression, StringEncoding stringEncoding);
 
 
-void fdsReadCharVec_v6(std::istream &myfile, IStringColumn* blockReader, unsigned long long blockPos, unsigned int startRow, unsigned int vecLength, unsigned int size);
+void fdsReadCharVec_v6(std::istream &myfile, IStringColumn* blockReader, unsigned long long blockPos, unsigned long long startRow,
+  unsigned long long vecLength, unsigned long long size);
 
 
 #endif  // CHARACTER_V6_H
