+++ src/fstcore/integer/integer_v8.cpp
@@ -41,7 +41,7 @@
 using namespace std;
 
 
-void fdsWriteIntVec_v8(ofstream &myfile, int* integerVector, unsigned int nrOfRows, unsigned int compression, std::string annotation)
+void fdsWriteIntVec_v8(ofstream &myfile, int* integerVector, unsigned long long nrOfRows, unsigned int compression, std::string annotation)
 {
   int blockSize = 4 * BLOCKSIZE_INT;  // block size in bytes
 
@@ -78,8 +78,8 @@ void fdsWriteIntVec_v8(ofstream &myfile, int* integerVector, unsigned int nrOfRo
 }
 
 
-void fdsReadIntVec_v8(istream &myfile, int* integerVec, unsigned long long blockPos, unsigned int startRow, unsigned int length,
-  unsigned int size, std::string &annotation)
+void fdsReadIntVec_v8(istream &myfile, int* integerVec, unsigned long long blockPos, unsigned long long startRow,
+  unsigned long long length, unsigned long long size, std::string &annotation)
 {
   return fdsReadColumn_v2(myfile, reinterpret_cast<char*>(integerVec), blockPos, startRow, length, size, 4, annotation, BATCH_SIZE_READ_INT);
 }
